+++ src/views/pages/Option/tabs/ScriptTab/index.vue
@@ -295,5 +295,3 @@ export default class ScriptTab extends Vue {
   }
 }
 </script>
-
-