+++ tests/match.test.ts
@@ -11,6 +11,8 @@ describe("UrlMatch", () => {
     url.add("https://bbs.tampermonkey.net.cn/*", "ok6")
     url.add("https://bbs.tampermonkey.net.cn/test/*", "ok66")
     url.add("*://*/test/param?*", "ok7")
+    url.add("i.tampermonkey.net.cn/*", "ok8")
+    url.add("*i.tampermonkey.net.cn/*", "ok9")
     it("match", () => {
         expect(url.match("https://www.baidu.com")).toEqual(["ok"]);
         expect(url.match("https://m.baidu.com")).toEqual(["ok"]);
@@ -30,10 +32,14 @@ describe("UrlMatch", () => {
         expect(url.match("https://bbs.tampermonkey.net.cn/test/thread-63-1-1.html")).toEqual(["ok6", "ok66"]);
         expect(url.match("https://bbs.tampermonkey.net.cn/forum-68-1.html")).toEqual(["ok6"]);
         expect(url.match("https://bbs.tampermonkey.net.cn/")).toEqual(["ok6"]);
-        expect(url.match("https:///bbs.tampermonkey.net.cn/test/param?a=1&b=2")).
+        expect(url.match("https://bbs.tampermonkey.net.cn/test/param?a=1&b=2")).
+            toEqual(["ok6", "ok66", "ok7"]);
+        expect(url.match("https://www.baidu.com/test/param")).
             toEqual(["ok7"]);
-        expect(url.match("https:///www.baidu.com/test/param")).
-            toEqual(["ok7"]);
+        expect(url.match("https://i.tampermonkey.net.cn/aa")).
+            toEqual(["ok8", "ok9"]);
+        expect(url.match("https://wwi.tampermonkey.net.cn/aa")).
+            toEqual(["ok9"]);
     });
     it("delete", () => {
         url.del("http://example.org/foo/bar.html", "ok5");